chore: refactor social media & topic template
