--- blocks/topic-takeaway/topic-takeaway.js
@@ -1,4 +1,0 @@
-export default function decorate(block) {
-  const main = document.querySelector('main');
-  main.prepend(block);
-}
