+++ templates/topic/topic.css
@@ -13,12 +13,12 @@
 .topic main .button-container a::after {
   background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='18' height='18' fill='currentColor' viewBox='0 0 18 18'%3E%3Cpath fill-rule='evenodd' d='M8.636 3.5a.5.5 0 0 0-.5-.5H1.5A1.5 1.5 0 0 0 0 4.5v10A1.5 1.5 0 0 0 1.5 16h10a1.5 1.5 0 0 0 1.5-1.5V7.864a.5.5 0 0 0-1 0V14.5a.5.5 0 0 1-.5.5h-10a.5.5 0 0 1-.5-.5v-10a.5.5 0 0 1 .5-.5h6.636a.5.5 0 0 0 .5-.5z'/%3E%3Cpath fill-rule='evenodd' d='M16 .5a.5.5 0 0 0-.5-.5h-5a.5.5 0 0 0 0 1h3.793L6.146 9.146a.5.5 0 1 0 .708.708L15 1.707V5.5a.5.5 0 0 0 1 0v-5z'/%3E%3C/svg%3E");
 }
- 
-.breadcrumb-container {
+
+.topic main > div:first-child {
   @apply space-y-4 mb-2;
 }
  
-.breadcrumb-container .default-content-wrapper h1 {
+.topic main > div:first-child .default-content-wrapper h1 {
   @apply text-gray-900 my-2 font-extrabold text-4xl py-2 px-0;    
 }
 
@@ -28,10 +28,6 @@
 
 *:has(.product-topic-takeaway) .default-content-wrapper h2 {
   @apply text-xl font-semibold text-danahergray-900 my-0;
-}
-
-.social-media-container h1 {
-  @apply text-gray-900 my-2 font-bold text-2xl mb-10 mt-20;
 }
 
 .topic main p:nth-of-type(2) {
@@ -59,12 +55,4 @@
 
 .topic main img {
   @apply w-full mt-10 mb-4;
-}
- 
-.ctasection-container .default-content-wrapper h1{
-  @apply text-gray-900 font-bold text-2xl;
-}
- 
-.ctasection-container .default-content-wrapper h2{
-  @apply inline-flex font-semibold text-xl text-danahergray-900;
 }
